pre-SLQ data call setup for local merge
I linked a new JS file with an empty function and added the basic HTML ground work for a call so I can just drop my function code in and test with on my laptop (not with me atm) and then push if/when its working.

Removed phantom php file from last commit.