Broken implement of loading screen WIP
@petercondoleon before you run around crazy looking for why the loading screen is broken, i realised now its becasue my document.body.contains() doesn't find the div element in the document for some reason.

maybe this is a silly implementation of handling the loading screen... I was tired when I did it. I have a timeout function checking if the "isloading" global variable is false, which is only suppose to be false when the ajax request is done??? idk maybe check the example to see if there is a more traight forward approach? maybe a function that just removes the loading screen in the success callback??

ill leave it with you for now